rework: Split exp tables hooks
